feat: add OneHandleSlider
added OneHandleSlider as a way to input data - currently used for distance. Additionally gave it a numeric value next to it etc. Basic styling, needed to add a new tw layer for it.
Also realized that it's better to make SettingsPage consist of separate components rather than one row of settings.
Did some refactoring for InputSelect and InputScroll.
Also added UserPreferences component to match the intention.