fix: signup/new user creation
when implementing new functionality it is a good practice to test if it still works with the added code
this is the exact reason why the unit tests are good, you don't have to go test it manually.
if you ran the unit tests you would have seen that they don't work.